Abstract expense form to separate component
